refactor: division 3 files & get labels config
